+++ go.mod
@@ -7,7 +7,6 @@ require (
 	github.com/ONSdigital/dp-healthcheck v1.0.5
 	github.com/ONSdigital/log.go/v2 v2.0.6
 	github.com/cucumber/godog v0.10.0
-	github.com/globalsign/mgo v0.0.0-20181015135952-eeefdecb41b8
 	github.com/smartystreets/goconvey v1.6.4
 	github.com/square/mongo-lock v0.0.0-20201208161834-4db518ed7fb2
 	github.com/stretchr/testify v1.6.1
